Migration to conan2 and remove wyrm dependency

+++ libcarma/include.h/carma_cublas.h
@@ -25,7 +25,7 @@
 #define carma_checkCublasStatus(status) \
   __carma_checkCublasStatus(status, __LINE__, __FILE__)
 
-cublasStatus_t __carma_checkCublasStatus(cublasStatus_t status, int line,
+cublasStatus_t __carma_checkCublasStatus(cublasStatus_t status, int32_t line,
                                          std::string file);
 
 cublasStatus_t carma_init_cublas(cublasHandle_t *cublas_handle);
@@ -43,106 +43,106 @@ cublasOperation_t carma_char2cublas_operation(char operation);
  */
 
 template <class T_data>
-int carma_where_amax(cublasHandle_t cublas_handle, int n, const T_data *vect,
-                     int incx);
+int32_t carma_where_amax(cublasHandle_t cublas_handle, int32_t n, const T_data *vect,
+                     int32_t incx);
 
 template <class T_data>
-int carma_where_amin(cublasHandle_t cublas_handle, int n, const T_data *vect,
-                     int incx);
+int32_t carma_where_amin(cublasHandle_t cublas_handle, int32_t n, const T_data *vect,
+                     int32_t incx);
 
 template <class T_data>
-T_data carma_getasum(cublasHandle_t cublas_handle, int n, const T_data *vect,
-                     int incx);
+T_data carma_getasum(cublasHandle_t cublas_handle, int32_t n, const T_data *vect,
+                     int32_t incx);
 
 template <class T_data>
-cublasStatus_t carma_axpy(cublasHandle_t cublas_handle, int n,
-                          const T_data alpha, const T_data *vectx, int incx,
-                          T_data *vecty, int incy);
+cublasStatus_t carma_axpy(cublasHandle_t cublas_handle, int32_t n,
+                          const T_data alpha, const T_data *vectx, int32_t incx,
+                          T_data *vecty, int32_t incy);
 
 template <class T_data>
-T_data carma_dot(cublasHandle_t cublas_handle, int n, T_data *vectx, int incx,
-                 T_data *vecty, int incy);
+T_data carma_dot(cublasHandle_t cublas_handle, int32_t n, T_data *vectx, int32_t incx,
+                 T_data *vecty, int32_t incy);
 
 template <class T_data>
-T_data carma_nrm2(cublasHandle_t cublas_handle, int n, T_data *vect, int incx);
+T_data carma_nrm2(cublasHandle_t cublas_handle, int32_t n, T_data *vect, int32_t incx);
 
 template <class T_data>
-cublasStatus_t carma_rot(cublasHandle_t cublas_handle, int n, T_data *vectx,
-                         int incx, T_data *vecty, int incy, T_data sc,
+cublasStatus_t carma_rot(cublasHandle_t cublas_handle, int32_t n, T_data *vectx,
+                         int32_t incx, T_data *vecty, int32_t incy, T_data sc,
                          T_data ss);
 
 template <class T_data>
-cublasStatus_t carma_scal(cublasHandle_t cublas_handle, int n, T_data alpha,
-                          T_data *vectx, int incx);
+cublasStatus_t carma_scal(cublasHandle_t cublas_handle, int32_t n, T_data alpha,
+                          T_data *vectx, int32_t incx);
 
 template <class T_data>
-cublasStatus_t carma_swap(cublasHandle_t cublas_handle, int n, T_data *vectx,
-                          int incx, T_data *vecty, int incy);
+cublasStatus_t carma_swap(cublasHandle_t cublas_handle, int32_t n, T_data *vectx,
+                          int32_t incx, T_data *vecty, int32_t incy);
 
 template <class T_data>
-cublasStatus_t carma_copy(cublasHandle_t cublas_handle, int n,
-                          const T_data *vectx, int incx, T_data *vecty,
-                          int incy);
+cublasStatus_t carma_copy(cublasHandle_t cublas_handle, int32_t n,
+                          const T_data *vectx, int32_t incx, T_data *vecty,
+                          int32_t incy);
 
 template <class T_data>
-cublasStatus_t carma_gemv(cublasHandle_t cublas_handle, char trans, int m,
-                          int n, T_data alpha, T_data *matA, int lda,
-                          T_data *vectx, int incx, T_data beta, T_data *vecty,
-                          int incy);
+cublasStatus_t carma_gemv(cublasHandle_t cublas_handle, char trans, int32_t m,
+                          int32_t n, T_data alpha, T_data *matA, int32_t lda,
+                          T_data *vectx, int32_t incx, T_data beta, T_data *vecty,
+                          int32_t incy);
 
 template <class T_data>
-cublasStatus_t carma_symv(cublasHandle_t cublas_handle, char uplo, int n,
-                          T_data alpha, T_data *matA, int lda, T_data *vectx,
-                          int incx, T_data beta, T_data *vecty, int incy);
+cublasStatus_t carma_symv(cublasHandle_t cublas_handle, char uplo, int32_t n,
+                          T_data alpha, T_data *matA, int32_t lda, T_data *vectx,
+                          int32_t incx, T_data beta, T_data *vecty, int32_t incy);
 
 template <class T_data>
-cublasStatus_t carma_ger(cublasHandle_t cublas_handle, int m, int n,
-                         T_data alpha, T_data *vectx, int incx, T_data *vecty,
-                         int incy, T_data *matA, int lda);
+cublasStatus_t carma_ger(cublasHandle_t cublas_handle, int32_t m, int32_t n,
+                         T_data alpha, T_data *vectx, int32_t incx, T_data *vecty,
+                         int32_t incy, T_data *matA, int32_t lda);
 
 template <class T_data>
 cublasStatus_t carma_gemm(cublasHandle_t cublas_handle, char transa,
-                          char transb, int m, int n, int k, T_data alpha,
-                          T_data *matA, int lda, T_data *matB, int ldb,
-                          T_data beta, T_data *matC, int ldc);
+                          char transb, int32_t m, int32_t n, int32_t k, T_data alpha,
+                          T_data *matA, int32_t lda, T_data *matB, int32_t ldb,
+                          T_data beta, T_data *matC, int32_t ldc);
 
 template <class T_data>
 cublasStatus_t carma_gemm_strided_batched(cublasHandle_t cublas_handle,
-                    char transa, char transb, int m, int n, int k, T_data alpha,
-                    T_data *matsA, int lda, long long int strideA,
-                    T_data *matsB, int ldb, long long int strideB,
-                    T_data beta, T_data *matsC, int ldc, long long int strideC,
-                    int batch_count);
+                    char transa, char transb, int32_t m, int32_t n, int32_t k, T_data alpha,
+                    T_data *matsA, int32_t lda, int64_t strideA,
+                    T_data *matsB, int32_t ldb, int64_t strideB,
+                    T_data beta, T_data *matsC, int32_t ldc, int64_t strideC,
+                    int32_t batch_count);
 
 template <class T_data>
 cublasStatus_t carma_symm(cublasHandle_t cublas_handle, char side, char uplo,
-                          int m, int n, T_data alpha, T_data *matA, int lda,
-                          T_data *matB, int ldb, T_data beta, T_data *matC,
-                          int ldc);
+                          int32_t m, int32_t n, T_data alpha, T_data *matA, int32_t lda,
+                          T_data *matB, int32_t ldb, T_data beta, T_data *matC,
+                          int32_t ldc);
 
 template <class T_data>
 cublasStatus_t carma_syrk(cublasHandle_t cublas_handle, char uplo, char transa,
-                          int n, int k, T_data alpha, T_data *matA, int lda,
-                          T_data beta, T_data *matC, int ldc);
+                          int32_t n, int32_t k, T_data alpha, T_data *matA, int32_t lda,
+                          T_data beta, T_data *matC, int32_t ldc);
 
 template <class T_data>
 cublasStatus_t carma_syrkx(cublasHandle_t cublas_handle, char uplo, char transa,
-                           int n, int k, T_data alpha, T_data *matA, int lda,
-                           T_data *matB, int ldb, T_data beta, T_data *matC,
-                           int ldc);
+                           int32_t n, int32_t k, T_data alpha, T_data *matA, int32_t lda,
+                           T_data *matB, int32_t ldb, T_data beta, T_data *matC,
+                           int32_t ldc);
 
 template <class T_data>
 cublasStatus_t carma_geam(cublasHandle_t cublas_handle, char transa,
-                          char transb, int m, int n, T_data alpha, T_data *matA,
-                          int lda, T_data beta, T_data *matB, int ldb,
-                          T_data *matC, int ldc);
+                          char transb, int32_t m, int32_t n, T_data alpha, T_data *matA,
+                          int32_t lda, T_data beta, T_data *matB, int32_t ldb,
+                          T_data *matC, int32_t ldc);
 
 template <class T_data>
-cublasStatus_t carma_dgmm(cublasHandle_t cublas_handle, char side, int m, int n,
-                          const T_data *matA, int lda, const T_data *vectx,
-                          int incx, T_data *matC, int ldc);
+cublasStatus_t carma_dgmm(cublasHandle_t cublas_handle, char side, int32_t m, int32_t n,
+                          const T_data *matA, int32_t lda, const T_data *vectx,
+                          int32_t incx, T_data *matC, int32_t ldc);
 
 template <class T_data>
-cublasStatus_t carma_sbmv(cublasHandle_t cublas_handle, char uplo, int N, int K, T_data alpha, T_data *A,
-                          int lda, T_data *x, int incx, T_data beta, T_data *y, int incy);
+cublasStatus_t carma_sbmv(cublasHandle_t cublas_handle, char uplo, int32_t N, int32_t K, T_data alpha, T_data *A,
+                          int32_t lda, T_data *x, int32_t incx, T_data beta, T_data *y, int32_t incy);
 #endif /* CARMA_CUBLAS_H_ */
